+++ packages/pwafire/src/index.ts
@@ -1,11 +1,49 @@
-/* Developer : Maye Edwin (Google Developer Expert):> Copyright : https://pwafire.org */
-import PWA from "./pwa";
+// Developer : Maye Edwin (Google Developer Expert):> Copyright : https://pwafire.org */
 import Check from "./check";
+import {
+  BadgingApi,
+  BarcodeDetectorApi,
+  ClipboardApi,
+  ConnectivityApi,
+  ContactsApi,
+  ContentIndexingApi,
+  FilesApi,
+  FontsApi,
+  FullscreenApi,
+  IdleDetectionApi,
+  InstallApi,
+  NotificationApi,
+  PaymentApi,
+  ShareApi,
+  VisibilityApi,
+  WakeLockApi,
+  WebOTPApi,
+} from "./pwa";
+
 const pwafire = {
-  pwa: new PWA(),
+  pwa: {
+    ...BadgingApi,
+    ...BarcodeDetectorApi,
+    ...ClipboardApi,
+    ...ConnectivityApi,
+    ...ContactsApi,
+    ...ContentIndexingApi,
+    ...FilesApi,
+    ...FontsApi,
+    ...FullscreenApi,
+    ...IdleDetectionApi,
+    ...InstallApi,
+    ...NotificationApi,
+    ...WebOTPApi,
+    ...PaymentApi,
+    ...ShareApi,
+    ...VisibilityApi,
+    ...WakeLockApi,
+  },
   check: new Check(),
 };
 
 const { pwa, check } = pwafire;
-export { pwa, check };
+export { check, pwa };
+
 export default pwafire;
