+++ packages/pwafire/src/pwa/fullscreen/index.ts
@@ -1,0 +1,14 @@
+export const FullscreenApi = {
+  Fullscreen: async () => {
+    try {
+      if (document.fullscreenEnabled) {
+        await document.documentElement.requestFullscreen();
+        return { ok: true, message: "Fullscreen" };
+      } else {
+        return { ok: false, message: "Fullscreen disabled" };
+      }
+    } catch (error) {
+      throw error;
+    }
+  },
+};
