+++ packages/pwafire/src/pwa/wake-lock/index.ts
@@ -1,0 +1,18 @@
+export const WakeLockApi = {
+  wakeLock: async () => {
+    try {
+      if ("wakeLock" in navigator) {
+        const wakeLock = await navigator.wakeLock.request("screen");
+        if (wakeLock) {
+          return { ok: true, message: "WakeLock Active" };
+        } else {
+          return { ok: false, message: "WakeLock Failed" };
+        }
+      } else {
+        return { ok: false, message: "WakeLock API not supported" };
+      }
+    } catch (error) {
+      throw error;
+    }
+  },
+};
